add required elasticsearch install for haystack
In order to use Elasticsearch with Django-Haystack, you need to have the
Python elasticsearch package installed in your virtual environment. See
the documentation here:

https://django-haystack.readthedocs.io/en/master/backend_support.html#id2

Without this, hitting the home page causes a 500 error with the message:

"MissingDependency: The 'elasticsearch' backend requires the
installation of 'elasticsearch'. Please refer to the documentation."