steps to make Python unit tests run on Travis
This commit makes a few changes to this repository to try to get the
Python unit tests working both locally and on Travis.

It sets up a `tox.ini` file for running Tox in two environments: one for
Django 1.8 and one for Django 1.11 (the two LTS versions). To run both,
just run `tox`, to run either, do e.g. `tox -e dj18` or `tox -e dj111`.
This uses its own virtual environment.

Currently the Django 1.8 environment fails because of a dependence on
`models.Index` which does not exist in that version. @grapesmoker are
you looking into fixing that?

There's an additional failure in
`eregs_core.eregs_tests.test_model_functions.ModelFunctionsTest.test_paragraph_functions`
due to some mismatched HTML; @grapesmoker can you also take a look at
that failure?

This commit adds a `.travis.yml` file for running automated unit tests
on TravisCI.

This commit adds an additional piece of test data into the `data`
subdirectory (which is now explicitly incorporated into
`eregs.settings`), because it is required to make one of the existing
unit tests pass. The wisdom of including this data in this repository is
perhaps something to be discussed further, but doing this continues an
existing pattern in the short-term interest of getting the tests running
for now.

The `setup.py` file has been modified to add some of the testing
dependencies required by Tox and Travis.

This commit also slightly modifies the `.gitignore` file to better hide
some testing-related files and directories.

Finally, this commit adds a `@unittest.skip` marker on some tests that
are hitting production servers; these shouldn't be part of the unit test
suite.