adds .coveragerc to properly report coverage
This change adds a `.coveragerc` file to properly report the unit test
code coverage. It omits certain directories that should not be counted
(`.tox`) and those that are somewhat unrelated to unit tests
(`migrations` and `settings`).

To see the difference this file makes, run unit tests with `tox` and
then type `coverage report` to generate the coverage report.