remove dependence on local_settings.py file
This change modifies how the `import_reg` management command works by
removing a dependence on a nonexistent `local_settings.py` file and
instead using an explicit command-line option, e.g.

./manage.py import_reg -r ../regulations-xml 1013

This makes the source of regulations more explicit and removes a setup
step when using this repository either manually or as part of an
automated process.

This change also modifies `import_diff` to use  `django.conf.settings`
instead of `eregs.local_settings` to check the `DEBUG` flag.

Finally, it also modifies the `bulk_create` call in `import_xml` to use
a batch size of 100; by default Django uses only a single query to
insert but if the amount of data to insert is excessive this can fail.
By setting a (hopefully reasonable) bulk batch size, this will avoid
errors when loading large regulations.