make migration eregs_core.0010 reversible
This commit make the `eregs_core.0010` migration reversible, to allow
for running Django migrations for this app in reverse. This is very
useful if you want to quickly reset the state of your DB, e.g. by doing

```sh
$ ./manage.py migrate eregs_core zero
$ ./manage.py migrate eregs_core
```

This resets the DB back to the state before any migrations were applied,
which is a useful shorthand for otherwise having to do something like

```sh
$ echo "DELETE FROM TABLE eregs_core_version;" | ./manage.py dbshell
$ echo "DELETE FROM TABLE eregs_core_regnode;" | ./manage.py dbshell
$ echo "DELETE FROM TABLE eregs_core_diffnode;" | ./manage.py dbshell
$ echo "DELETE FROM TABLE django_migrations WHERE app = 'eregs_core'; |
./manage.py dbshell"
```

This change just adds some reverse SQL to the manual migration to delete
the indexes that are added going the other way.