move all templates under eregs_core subdirectory
This change moves all app templates under an `eregs_core` subdirectory
within the `templates` directory to better conform with Django best
practices.

As mentioned in the Django tutorial

(https://docs.djangoproject.com/en/1.8/intro/tutorial03/#write-views-that-actually-do-something)

"Django will choose the first template it finds whose name matches, and
if you had a template with the same name in a different application,
Django would be unable to distinguish between them. We need to be able
to point Django at the right one, and the easiest way to ensure this is
by namespacing them. That is, by putting those templates inside another
directory named for the application itself."

This change also adds a few unit tests to verify that template rendering
works properly for the different views. Not all views are testable at
the moment but this is a reasonable start.

It also removes a few `print` statements that were in a couple of the
views.