cleanup setup.py, install only eregs_core
This change modifies how `setup.py` works and how this package gets
installed via `pip install`. Instead of installing two separate packages
(`eregs_core`, containing views and models, and `eregs`, containing URLs
and settings), this app now only installs a single `eregs_core` package
that contains models, views, and URLs. This makes the code more
self-contained and is more consistent with how users of this package
might expect it to be used.

This change also modifies the author and license in `setup.py` to be
more consistent with CFPB standards (using CC0 instead of MIT). It also
enables the use of setuptools-git-version to more easily set package
versions for releases.

Note that this will require a change to how projects like cfgov-refresh
incorporate this project's URLs.

This addresses issues #46 and #47.