fixes static asset build and serve
This commit fixes how the static built JS files are served by Django.
When running in development mode (with `./manage.py runserver`) we want
to serve the static JS from the `/static` subfolder of this app, which
is where the build puts its output.

In addition, this commit removes three third-party libraries from this
repo (jQuery, Bootstrap, and Modernizr) in favor of including them in
the `package.json` file for installation via `npm`.

To build static assets, the process in `README` is still correct: a
one-time build can be invoked via `npm run build`.

I've verified that this works if you do a `pip install` of this branch
into cfgov-refresh, although I'm not sure if the changes to how the
libraries are included might cause any conflicts. Tagging @contolini and
@sebworks for review on that part.

Additionally, this commit renames one of the templatetags files from
`processing.py` to `eregs_processing.py` to avoid conflict with another
`processing.py` apparently in cfgov-refresh; this just requires doing a
`{% load eregs_processing %}` in templates instead.