disable javascript
This commit disables JavaScript on this app's pages by removing
(commenting out) the inclusion of the JS library.

While this app is in pre-release mode and is still being developed,
removing the use of JS will help to simplify testing. The app should
also support non-JS users as well, so removing JS for now will help to
guarantee that all functionality works when JS is disabled.