convert hardcoded urls to dynamic ones
This change removes usage of hardcoded URLs in favor of using Django's
`django.core.urlresolvers.reverse` and `{% url %}` template tag to
better support cases where this app is being deployed under a website's
subdirectory.

For example, if this app is deployed under a `/eregs-2.0` subdirectory,
regulation URLs should be under `/eregs-2.0/regulation/etc`, and not
under a hardcoded `/regulation/etc`.

In addition, a couple of hardcoded `/static` references in templates
have been converted to use of Django's `{% static %}` tag, for a similar
reason.

All existing unit tests continue to pass, and working site functionality
continues to do so.