cleanup main view and add tests
This commit converts the landing page (main view) into a Django
class-based view and moves it into its own module. It also adds various
unit tests to cover this view.

This is a partial fix to the issue in #44 where non-GET (e.g. POST)
requests to this view would raise a 500 error.

This commit adds a dummy `data/header-only.xml` RegML file (which
validates against the schema) that can be used for simple testing of
this view only. The new unit tests for this view validate that it works
with no data or with data loaded.

This commit also adds a simple unittest against `eregs_core.urls` to
ensure that the URLs module can be imported (to protect against syntax
errors which would otherwise not be discovered via the test suite).